Remove restoreBlocks from synchronizer and move to utils
